requirements.txt
Pillow
rich
Flask
werkzeug